moved classify and classify_string_classifier_rule and their tests from reader to classifier_services and classifier_services_test
